requirements.txt
implicit==0.4.0
pandas==0.25.3
scipy==1.2.0
python-dotenv
numpy==1.16.0
Flask
gunicorn
sklearn
